Add rage counter to city

+++ Assets/Scripts/Rage.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 
 public class Rage
 {
-    public static int value = 0;
+    public static float value = 0;
     public static int MaxRage = 2000;
 }
